pyproject.toml
[project]
name = "pi-gate"
version = "0.1.0"
description = ""
authors = [
    {name = "shubham",email = "sk246903@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "dash (>=2.18.2,<3.0.0)",
    "dnslib (>=0.9.25,<0.10.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "aiohttp (>=3.11.12,<4.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "uvloop (>=0.21.0,<0.22.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pybloom-live (>=4.0.0,<5.0.0)",
]
packages = [{ include = "pi_gate" }]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[project.scripts]
pi-gate = "pi_gate.app:app"

